[NNVM][FRONTEND] Review comments.
