+++ tutorials/nnvm/from_tensorflow.py
@@ -14,7 +14,6 @@ import nnvm
 import tvm
 import numpy as np
 import os.path
-import re
 
 # Tensorflow imports
 import tensorflow as tf
@@ -22,8 +21,9 @@ from tensorflow.core.framework import graph_pb2
 from tensorflow.python.framework import dtypes
 from tensorflow.python.framework import tensor_util
 
+import nnvm.testing.tf
 
-repo_base = 'https://github.com/srkreddy1238/dmlc_data/raw/master/models/tensorflow/InceptionV1/'
+repo_base = 'https://github.com/dmlc/web-data/raw/master/tensorflow/models/InceptionV1/'
 img_name = 'elephant-299.jpg'
 image_url = os.path.join(repo_base, img_name)
 model_name = 'classify_image_graph_def-with_shapes.pb'
@@ -35,88 +35,8 @@ lable_map_url = os.path.join(repo_base, lable_map)
 
 
 ######################################################################
-# Some helper functions
-
-def _ProcessGraphDefParam(graph_def):
-    """Type-checks and possibly canonicalizes `graph_def`."""
-    if not isinstance(graph_def, graph_pb2.GraphDef):
-        # `graph_def` could be a dynamically-created message, so try a duck-typed
-        # approach
-        try:
-            old_graph_def = graph_def
-            graph_def = graph_pb2.GraphDef()
-            graph_def.MergeFrom(old_graph_def)
-        except TypeError:
-            raise TypeError('graph_def must be a GraphDef proto.')
-    return graph_def
-
-class NodeLookup(object):
-    """Converts integer node ID's to human readable labels."""
-
-    def __init__(self,
-                 label_lookup_path=None,
-                 uid_lookup_path=None):
-        if not label_lookup_path:
-            label_lookup_path = os.path.join(
-                "./", map_proto)
-        if not uid_lookup_path:
-            uid_lookup_path = os.path.join(
-                "./", lable_map)
-        self.node_lookup = self.load(label_lookup_path, uid_lookup_path)
-
-    def load(self, label_lookup_path, uid_lookup_path):
-        """Loads a human readable English name for each softmax node.
-
-        Args:
-            label_lookup_path: string UID to integer node ID.
-            uid_lookup_path: string UID to human-readable string.
-
-        Returns:
-            dict from integer node ID to human-readable string.
-        """
-        if not tf.gfile.Exists(uid_lookup_path):
-            tf.logging.fatal('File does not exist %s', uid_lookup_path)
-        if not tf.gfile.Exists(label_lookup_path):
-            tf.logging.fatal('File does not exist %s', label_lookup_path)
-
-        # Loads mapping from string UID to human-readable string
-        proto_as_ascii_lines = tf.gfile.GFile(uid_lookup_path).readlines()
-        uid_to_human = {}
-        p = re.compile(r'[n\d]*[ \S,]*')
-        for line in proto_as_ascii_lines:
-            parsed_items = p.findall(line)
-            uid = parsed_items[0]
-            human_string = parsed_items[2]
-            uid_to_human[uid] = human_string
-
-        # Loads mapping from string UID to integer node ID.
-        node_id_to_uid = {}
-        proto_as_ascii = tf.gfile.GFile(label_lookup_path).readlines()
-        for line in proto_as_ascii:
-            if line.startswith('  target_class:'):
-                target_class = int(line.split(': ')[1])
-            if line.startswith('  target_class_string:'):
-                target_class_string = line.split(': ')[1]
-                node_id_to_uid[target_class] = target_class_string[1:-2]
-
-        # Loads the final mapping of integer node ID to human-readable string
-        node_id_to_name = {}
-        for key, val in node_id_to_uid.items():
-            if val not in uid_to_human:
-                tf.logging.fatal('Failed to locate: %s', val)
-            name = uid_to_human[val]
-            node_id_to_name[key] = name
-
-        return node_id_to_name
-
-    def id_to_string(self, node_id):
-        if node_id not in self.node_lookup:
-            return ''
-        return self.node_lookup[node_id]
-
-######################################################################
 # Download processed tensorflow model
-# ---------------------------------------------
+# -----------------------------------
 # In this section, we download a pretrained Tensorflow model and classify an image.
 from mxnet.gluon.utils import download
 
@@ -128,6 +48,34 @@ download(lable_map_url, lable_map)
 
 ######################################################################
 # Creates graph from saved graph_def.pb.
+# --------------------------------------
+
+def _ProcessGraphDefParam(graph_def):
+    """Type-checks and possibly canonicalizes `graph_def`.
+
+    Parameters
+    ----------
+    graph_def : Obj
+        tensorflow graph definition.
+
+    Returns
+    -------
+    graph_def : Obj
+        tensorflow graph devinition
+
+    """
+
+    if not isinstance(graph_def, graph_pb2.GraphDef):
+        # `graph_def` could be a dynamically-created message, so try a duck-typed
+        # approach
+        try:
+            old_graph_def = graph_def
+            graph_def = graph_pb2.GraphDef()
+            graph_def.MergeFrom(old_graph_def)
+        except TypeError:
+            raise TypeError('graph_def must be a GraphDef proto.')
+    return graph_def
+
 with tf.gfile.FastGFile(os.path.join(
         "./", model_name), 'rb') as f:
     graph_def = tf.GraphDef()
@@ -138,6 +86,7 @@ with tf.gfile.FastGFile(os.path.join(
 
 ######################################################################
 # Decode image
+# ------------
 from PIL import Image
 image = Image.open(img_name).resize((299, 299))
 
@@ -150,7 +99,7 @@ print('x', x.shape)
 
 ######################################################################
 # Import the graph to NNVM
-# -----------------
+# ------------------------
 sym, params = nnvm.frontend.from_tensorflow(graph_def)
 
 ######################################################################
@@ -165,6 +114,7 @@ graph, lib, params = nnvm.compiler.build(sym, target, shape_dict, dtype=dtype_di
 
 ######################################################################
 # Save the compilation output.
+# ----------------------------
 """
 lib.export_library("imagenet_tensorflow.so")
 with open("imagenet_tensorflow.json", "w") as fo:
@@ -197,7 +147,8 @@ predictions = tvm_output.asnumpy()
 predictions = np.squeeze(predictions)
 
 # Creates node ID --> English string lookup.
-node_lookup = NodeLookup()
+node_lookup = nnvm.testing.tf.NodeLookup(label_lookup_path=os.path.join("./", map_proto),
+                                         uid_lookup_path=os.path.join("./", lable_map))
 
 top_k = predictions.argsort()[-5:][::-1]
 for node_id in top_k:
@@ -221,10 +172,13 @@ def create_graph():
 def run_inference_on_image(image):
     """Runs inference on an image.
 
-    Args:
-        image: Image file name.
+    Parameters
+    ----------
+    image: String
+        Image file name.
 
-    Returns:
+    Returns
+    -------
         Nothing
     """
     if not tf.gfile.Exists(image):
@@ -242,7 +196,8 @@ def run_inference_on_image(image):
         predictions = np.squeeze(predictions)
 
         # Creates node ID --> English string lookup.
-        node_lookup = NodeLookup()
+        node_lookup = nnvm.testing.tf.NodeLookup(label_lookup_path=os.path.join("./", map_proto),
+                                                 uid_lookup_path=os.path.join("./", lable_map))
 
         top_k = predictions.argsort()[-5:][::-1]
         print ("===== TENSORFLOW RESULTS =======")
